Refactor sitemap generation and remove unused code
The sitemap.ts file was refactored to improve performance and remove unused code. The code for fetching portals and prizes was removed as it was no longer needed. Additionally, the sitemap now only includes the about page for the portal. This change improves the efficiency of the sitemap generation process.

Fix incorrect address in portalAdminCard component

The portalAdminCard.tsx component had an incorrect address for a contract. This commit fixes the address to ensure that the correct contract is used.