refactor(infra): update imageBucket initialization
The imageBucket initialization in the storage module has been refactored to use the sst.aws.Bucket.get() method instead of conditionally creating a new Bucket instance. This change simplifies the code and improves consistency.

Also, commented out the router import in sst.config.ts as it is currently not being used.